+++ test/utils/connect.ts
@@ -8,13 +8,9 @@ interface ExtraConnectionConfig {
 }
 
 // to not duplicate code
-const staticOptions = {
-  useNewUrlParser: true,
-  useFindAndModify: true,
-  useCreateIndex: true,
-  useUnifiedTopology: true,
+const staticOptions: mongoose.ConnectOptions = {
   autoIndex: true,
-} as mongoose.ConnectionOptions;
+};
 
 /**
  * Make a Connection to MongoDB
@@ -39,7 +35,7 @@ export async function connect(extraConfig: ExtraConnectionConfig = {}): Promise<
   const connectionString = `${process.env.MONGO_URI}/${extraConfig.dbName ?? config.DataBase}`;
 
   if (extraConfig.createNewConnection) {
-    connection = mongooseInstance.createConnection(connectionString, options);
+    connection = await mongooseInstance.createConnection(connectionString, options).asPromise();
   } else {
     await mongoose.connect(connectionString, options);
     connection = mongooseInstance.connection;